requirements.txt
aiogram>=2.11.2 
python-dotenv>=0.15.0